requirements.txt
fastapi>=0.115.0
uvicorn==0.24.0
pydantic>=2.9.0
opik==1.9.98
python-dotenv==1.0.0
sqlalchemy
openai
google-genai
groq
pytest
